wip: rename legacy decorator
