+++ src/application/create-application.ts
@@ -4,11 +4,11 @@ import { serveStaticFiles } from '@/Http/Files';
 import { applyConfiguredGlobalMiddleware } from '@/Http/middleware/apply-configured-global-middleware';
 import { initializeRequestScopeStorage } from '@/Http/Scope/request-scope-storage';
 import { registerEventSubscribers } from '@/Kernel';
-import { registerLegacyRoutes } from '@/routing/decorator/legacy-router';
+import { registerLegacyRoutes } from '@/routing/legacy-decorator/legacy-router';
 import { registerRoutes } from '@/routing/registration/register-routes';
 import { verifyRoutingMode } from '@/routing/verify-routing-mode';
-import Koa from 'koa';
 import type { Request } from 'koa';
+import Koa from 'koa';
 import { type Application } from './application';
 
 export function create<TRequest extends Request>(config: KoalaConfig<TRequest>): Application {
